+++ shared/rocroller/lib/include/rocRoller/KernelGraph/Transforms/OrderExchangeNodes.hpp
@@ -32,27 +32,15 @@ namespace rocRoller
 {
     namespace KernelGraph
     {
-        ConstraintStatus NoUnorderedMultiplyNodes(const KernelGraph& k);
-
-        class OrderMultiplyNodes : public GraphTransform
+        class OrderExchangeNodes : public GraphTransform
         {
         public:
-            OrderMultiplyNodes() = default;
+            OrderExchangeNodes() = default;
 
             KernelGraph apply(KernelGraph const& original) override;
             std::string name() const override
             {
-                return "OrderMultiplyNodes";
-            }
-
-            inline std::vector<GraphConstraint> preConstraints() const override
-            {
-                return {};
-            }
-
-            inline std::vector<GraphConstraint> postConstraints() const override
-            {
-                return {&NoUnorderedMultiplyNodes};
+                return "OrderExchangeNodes";
             }
         };
     }
