+++ shared/rocroller/lib/source/KernelGraph/Transformations/CMakeLists.txt
@@ -31,11 +31,13 @@ target_sources(rocroller
         "${CMAKE_CURRENT_SOURCE_DIR}/LowerTile.cpp"
         "${CMAKE_CURRENT_SOURCE_DIR}/NopExtraScopes.cpp"
         "${CMAKE_CURRENT_SOURCE_DIR}/OrderEpilogueBlocks.cpp"
+        "${CMAKE_CURRENT_SOURCE_DIR}/OrderExchangeNodes.cpp"
         "${CMAKE_CURRENT_SOURCE_DIR}/OrderMemory.cpp"
         "${CMAKE_CURRENT_SOURCE_DIR}/OrderMultiplyNodes.cpp"
         "${CMAKE_CURRENT_SOURCE_DIR}/PrefetchScale.cpp"
         "${CMAKE_CURRENT_SOURCE_DIR}/RemapOutputTiles.cpp"
         "${CMAKE_CURRENT_SOURCE_DIR}/RemoveDuplicates.cpp"
+        "${CMAKE_CURRENT_SOURCE_DIR}/RemoveImplicitScheduling.cpp"
         "${CMAKE_CURRENT_SOURCE_DIR}/RemoveSetCoordinate.cpp"
         "${CMAKE_CURRENT_SOURCE_DIR}/ScheduleMultiplyAndLDS.cpp"
         "${CMAKE_CURRENT_SOURCE_DIR}/Simplify.cpp"
